add zod and react-hook-form

+++ src/actions/pets.ts
@@ -3,12 +3,20 @@
 import { revalidatePath } from "next/cache";
 import { Pet } from "@prisma/client";
 import prisma from "@/lib/db";
+import { petFormSchema, petIdSchema } from "@/schemas/pet";
 
 export const addPetAction = async (
   pet: Omit<Pet, "id" | "updatedAt" | "createdAt">
 ) => {
+  const { success, data } = petFormSchema.safeParse(pet);
+
+  if (!success) {
+    return {
+      message: "Invalid pet data",
+    };
+  }
   try {
-    await prisma.pet.create({ data: pet });
+    await prisma.pet.create({ data });
   } catch (error) {
     return {
       message: "Could not add pet",
@@ -22,12 +30,22 @@ export const editPetAction = async (
   id: string,
   pet: Omit<Pet, "id" | "updatedAt" | "createdAt">
 ) => {
+  const { success: petSuccess, data: validatedPet } =
+    petFormSchema.safeParse(pet);
+  const { data: validatedId, success: idSuccess } = petIdSchema.safeParse(id);
+
+  if (!petSuccess || !idSuccess) {
+    return {
+      message: "Invalid pet data",
+    };
+  }
+
   try {
     await prisma.pet.update({
       where: {
-        id,
+        id: validatedId,
       },
-      data: pet,
+      data: validatedPet,
     });
   } catch (error) {
     return {
@@ -39,10 +57,18 @@ export const editPetAction = async (
 };
 
 export const deletePetAction = async (id: string) => {
+  const { data, success } = petIdSchema.safeParse(id);
+
+  if (!success) {
+    return {
+      message: "Invalid pet id",
+    };
+  }
+
   try {
     await prisma.pet.delete({
       where: {
-        id,
+        id: data,
       },
     });
   } catch (error) {
